feat: Refactor image service and optimize image saving
Refactor the image service to optimize image resizing and conversion. Remove the `hostUrl` parameter from the `saveImage` function as it is no longer needed. Update the `createShorter` function in the `shorter` service to remove the `hostUrl` parameter and use the `slug` as the short URL. This commit improves the efficiency and simplicity of the image service and shortening functionality.